requirements.txt
Flask
opencv-contrib-python
numpy
pandas
flask-login
flask-mail
Pillow
APScheduler
cryptography
gunicorn
psycopg2-binary
